cli/composefs: Handle rollback for composefs
This commit does not handle the edge cases that arise with rollbacks
when there is a staged deployment present.

For Grub BLS case, we create a new `loader/entries.staged` directory,
write the new boot entries, then atomically swap `loader/entries` and
`loader/entries.staged`.

For Grub UKI case, we regenerate `grub2/user.cfg` using the images
present in `/sysroot/state/deploy/`

To distinguish whether the currently booted system is booted with a UKI
or BLS, we add an entry to origin file called `boot_type`

Signed-off-by: Johan-Liebert1 <pragyanpoudyal41999@gmail.com>